feat: ERKAP - Implement Financial Projection & Approval Workflow

+++ app/Http/Controllers/Erkap/RiskIdentificationController.php
@@ -9,6 +9,7 @@ use App\Models\Erkap\DepartmentTarget;
 use App\Models\Erkap\RiskIdentification;
 use App\Models\Erkap\RiskTaxonomy;
 use App\Models\Erkap\RiskType;
+use App\Services\ErkapAccess;
 use Exception;
 
 class RiskIdentificationController extends Controller
@@ -16,7 +17,12 @@ class RiskIdentificationController extends Controller
     public function index()
     {
         $pageName = 'Identifikasi Risiko';
-        $riskIdentifications = RiskIdentification::with(['departmentTarget', 'riskType', 'riskTaxonomy'])->paginate(10);
+        $riskIdentifications = RiskIdentification::with(['departmentTarget.ratingCriteria', 'riskType', 'riskTaxonomy'])
+            ->withCount('workPrograms')
+            ->when(ErkapAccess::isDivisionScoped(), function ($query) {
+                $query->whereIn('erkap_department_target_id', ErkapAccess::departmentTargetIds());
+            })
+            ->paginate(10);
 
         return view('erkap.risk-identification.index', compact('pageName', 'riskIdentifications'));
     }
@@ -24,7 +30,11 @@ class RiskIdentificationController extends Controller
     public function create()
     {
         $pageName = 'Buat Identifikasi Risiko';
-        $departmentTargets = DepartmentTarget::all();
+        $departmentTargets = DepartmentTarget::query()
+            ->when(ErkapAccess::isDivisionScoped(), function ($query) {
+                $query->where('division_id', ErkapAccess::divisionId());
+            })
+            ->get();
         $riskTypes = RiskType::all();
         $riskTaxonomies = RiskTaxonomy::all();
 
@@ -34,6 +44,8 @@ class RiskIdentificationController extends Controller
     public function store(StoreRiskIdentificationRequest $request)
     {
         try {
+            ErkapAccess::assertDepartmentTargetAccess($request->integer('erkap_department_target_id'));
+
             RiskIdentification::create($request->validated());
 
             return redirect()->route('erkap.risk-identifications.index')
@@ -52,8 +64,14 @@ class RiskIdentificationController extends Controller
 
     public function edit(RiskIdentification $riskIdentification)
     {
+        ErkapAccess::assertDepartmentTargetAccess($riskIdentification->erkap_department_target_id);
+
         $pageName = 'Edit Identifikasi Risiko';
-        $departmentTargets = DepartmentTarget::all();
+        $departmentTargets = DepartmentTarget::query()
+            ->when(ErkapAccess::isDivisionScoped(), function ($query) {
+                $query->where('division_id', ErkapAccess::divisionId());
+            })
+            ->get();
         $riskTypes = RiskType::all();
         $riskTaxonomies = RiskTaxonomy::all();
 
@@ -63,6 +81,9 @@ class RiskIdentificationController extends Controller
     public function update(UpdateRiskIdentificationRequest $request, RiskIdentification $riskIdentification)
     {
         try {
+            ErkapAccess::assertDepartmentTargetAccess($riskIdentification->erkap_department_target_id);
+            ErkapAccess::assertDepartmentTargetAccess($request->integer('erkap_department_target_id'));
+
             $riskIdentification->update($request->validated());
 
             return redirect()->route('erkap.risk-identifications.index')
@@ -82,6 +103,13 @@ class RiskIdentificationController extends Controller
     public function destroy(RiskIdentification $riskIdentification)
     {
         try {
+            ErkapAccess::assertDepartmentTargetAccess($riskIdentification->erkap_department_target_id);
+
+            if ($riskIdentification->hasWorkProgram()) {
+                return redirect()->route('erkap.risk-identifications.index')
+                    ->with('error', 'Risiko yang sudah memiliki program kerja tidak bisa dihapus.');
+            }
+
             $riskIdentification->delete();
 
             return redirect()->route('erkap.risk-identifications.index')
